Rename export plugins helpers
